Electric shield animation
These are the sprites for the shield power-up animation. The sprites labeled "top" should be put on top of the player airship sprite when the shield is activated. The ones labeled "bottom" should be put behind the player sprite. The optimal in-engine playback of the animation would be somewhere between 6 and 10 fps.